Add CapacityExpansion.jl
Add CapacityExpansion.jl in Energy System Assessment - Modeling. 

The package allows for an expandable modeling of generation and transmission capacity expansion planning for future energy systems. It provides both simple-to-use code and sample datasets. 

The package complies with the guideline because it is open source both in its software and provides some sample datasets, has been published in the journal of open source software, has been contributed to by authors beyond the core authors (see issues/PRs), is structured so that it can be reused and extended (including that we provide comments and documentation).