Add Brick
HVAC systems accout for most of energy usage in buildings.  Currently, the HVAC systems data is usually defined in a messy way. This package is developed based on RDF model, allowing a uniformed and efficient way to query the data point.

I think it aligns to sustainabilty, because it construct the basic data engineering of HVAC systems and building digitalization. Using AI to imporve the building energy would be impossibel without this infrastructure.